adjust examples' file structures; change test-code style by adding "if __name__ == "__main__":"
