--- examples/__init__.py
@@ -1,2 +1,0 @@
-__all__ = ['pyalgo',
-           'data_structure', 'data_structure_Queue', 'data_structure_Stack']
